del:the level of nothing
